fix(qqmusic): 取流改走侧边栏音乐卡片同款 Lite 限免链路(ztelite comm + ppurl + CgiGetTempVkey)；修 companion/Int-Float 编译错误；symphonia 开 vorbis 解码 ogg

+++ app/src/main/java/dev/ujhhgtg/wekit/features/items/chat/ChatAnalysisPng.kt
@@ -575,15 +575,15 @@ object ChatAnalysisPng {
         val metaP = paint(FS_META, COLOR_META)
         val smallP = paint(FS_SMALL, COLOR_META)
         drawTextLines(
-            cv, spec.nameLines, spec.textLeft.toFloat(), inner + spec.nameTop,
+            cv, spec.nameLines, spec.textLeft.toFloat(), (inner + spec.nameTop).toFloat(),
             HEADER_TITLE_LINE_H, titleP, spec.textRight.toFloat(), innerBottom,
         )
         drawTextLines(
-            cv, spec.subLines, spec.textLeft.toFloat(), inner + spec.subTop,
+            cv, spec.subLines, spec.textLeft.toFloat(), (inner + spec.subTop).toFloat(),
             HEADER_META_LINE_H, metaP, spec.textRight.toFloat(), innerBottom,
         )
         drawTextLines(
-            cv, spec.genLines, spec.textLeft.toFloat(), inner + spec.genTop,
+            cv, spec.genLines, spec.textLeft.toFloat(), (inner + spec.genTop).toFloat(),
             HEADER_META_LINE_H, smallP, spec.textRight.toFloat(), innerBottom,
         )
     }
